[DEVELOPMENT ENVIRONMENT] * BUG: Try to fix the spurious error that is raised sometimes when clicking   on a control in the form editor.
[GB.FORM]
* NEW: GridView.Columns.Width property has been implemented.


git-svn-id: svn://localhost/gambas/trunk@4519 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_alloc.h
@@ -48,7 +48,7 @@ typedef
   PACKED
   ALLOC;
 
-EXTERN int MEMORY_size;
+EXTERN size_t MEMORY_size;
 EXTERN FILE *MEMORY_log;
 
 #define ALLOC(_ptr, _size, _src)        MEMORY_alloc((void *)_ptr, _size, _src)
